ENH: Implement erfcx for float, double and long double.
The code for erfcx is now generated by a Python script that
uses mpmath.  The generated code is modeled after the old version
of erfcx, as written by Steven G. Johnson, but the new version has
a function written specifically for each floating point type.  This
allows a higher precision result with `long double`, and a faster
result with `float`.

To avoid making mpmath a build dependency, the generated code is
committed in src/normal, along with the Python script that generates
the code, `generate_erfcx_code.py`.