ENH: Add tools for creating gufuncs from C++ templated core functions.
Use the new tools to implement vnorm.  The source for vnorm now consists
of a C++ header file, vnorm_gufunc.h, that defines two templated C++
functions that implement the core calculation for real and complex
types, and a configuration file, define_cxx_gufunc_extmod.py, that
gives the information needed to generate an extension module that
implements the gufunc.